backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0
aiohttp==3.11.9
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
pytz==2024.2
python-dateutil==2.9.0
resend==2.4.0
cachetools==5.5.0